Move code into module package.
