+++ mysensors/__init__.py
@@ -1,0 +1,3 @@
+"""
+Python implementation of MySensors API
+"""
