+++ mysensors/const.py
